+++ src/components/common/Header/Header.tsx
@@ -6,73 +6,18 @@ import PersonIcon from '@mui/icons-material/Person';
 import LogoutIcon from '@mui/icons-material/Logout';
 
 import { useEffect, useState } from 'react';
-import { StyledMenu, StyledMenuItem } from './StyledMenu';
+import { StyledMenu, StyledMenuItem } from '../StyledMenu';
 
-import { ReactComponent as LogoMainIcon } from '../../assets/icons/logo-main.svg';
-import { ReactComponent as LogoSubIcon } from '../../assets/icons/logo-hover.svg';
-import { ReactComponent as CancleIcon } from '../../assets/icons/cancle.svg';
+import { ReactComponent as LogoMainIcon } from '../../../assets/icons/logo-main.svg';
+import { ReactComponent as CancleIcon } from '../../../assets/icons/cancle.svg';
 
-import { useAuth } from '../../AuthenticationContext';
-import { typography } from '../../constants/themeValue';
-import BASE_URL from '../../config';
-import fetchNotification from '../../api/notification';
-import { Notification } from '../../types/notification';
-import TimeSincePost from '../album/TimeSincePost';
-
-function NotificationContents({ noti }) {
-  return (
-    <Box sx={{ flexDirection: 'column', margin: '8px 0px' }}>
-      <Typography
-        fontWeight="light"
-        fontSize={typography.size.md}
-        sx={{
-          color: 'rgb(126, 126, 126)',
-          mb: '4px',
-        }}
-      >
-        <TimeSincePost createdAt={noti?.timestamp} />
-      </Typography>
-      <Typography>
-        {noti?.sender_id?.name}님이&nbsp;
-        {noti?.review_id?.title}에&nbsp;
-        {noti?.type === '좋아요' ? `${noti?.type} 표시를 했습니다.` : `${noti?.type}을 달았습니다.`}
-      </Typography>
-      <Typography fontSize={typography.size.md} sx={{ mt: '4px' }}>
-        {noti?.text}
-      </Typography>
-    </Box>
-  );
-}
-
-function LogoHoverIcon() {
-  const [count, setCount] = useState(false);
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      setCount(!count);
-    }, 500);
-
-    return () => {
-      clearInterval(timer);
-    };
-  });
-
-  return count ? (
-    <LogoMainIcon
-      style={{
-        width: 125,
-        height: 20,
-      }}
-    />
-  ) : (
-    <LogoSubIcon
-      style={{
-        width: 125,
-        height: 20,
-      }}
-    />
-  );
-}
+import { useAuth } from '../../../AuthenticationContext';
+import { typography } from '../../../constants/themeValue';
+import BASE_URL from '../../../config';
+import fetchNotification from '../../../api/notification';
+import { Notification } from '../../../types/notification';
+import NotificationContents from './NotificationContents';
+import LogoHoverIcon from './LogoHoverIcon';
 
 export default function MenuAppBar() {
   const [notifications, setNotifications] = useState<Notification[]>([]);
@@ -121,6 +66,8 @@ export default function MenuAppBar() {
   const handleCloseNoti = () => {
     setAnchorNoti(null);
   };
+
+  console.log(notifications);
 
   return (
     <Box>
